requirements/prod.txt
-r common.txt

boto==2.1.1

cssmin==0.1.4

django-compressor==1.1.2

django-htmlmin==0.5.1

django-pylibmc-sasl==0.2.4

django-storages==1.1.3

gunicorn==0.14.1

newrelic==1.2.0.246

psycopg2==2.4.5

pylibmc==1.2.2

raven==1.3.5

slimit==0.6